requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow==2.14.0
numpy==1.24.3
gunicorn==21.2.0

